Enhancement: (Re)add support for float timeout and timestamp values

+++ src/dNG/pas/data/tasks/database.py
@@ -152,7 +152,7 @@ Returns the task for the given TID.
 		"""
 Get the implementation specific next "run()" UNIX timestamp.
 
-:return: (int) UNIX timestamp; -1 if no further "run()" is required at the
+:return: (float) UNIX timestamp; -1 if no further "run()" is required at the
          moment
 :since:  v0.1.00
 		"""
